OIDC Username Claim
The Kubernetes API prefixes the 'IssuerURL' to all UsernameClaims other
than email.  To use the nickname claim, we must prefix the subject's name with the IssuerURL until which point kops supports
the `--oidc-username-prefix` flag. See: https://kubernetes.io/docs/admin/authentication/#configuring-the-api-server

Also see Aldo's PR where he added support for `--oidc-username-prefix`
to kops: https://github.com/kubernetes/kops/pull/4085

Modifying `kind` in roleRef as the admin role is a clusterRole which we
are scoping to a namespace.