kube2iam: add toleration for highmem nodes
new `highmem` instanceGroups are tainted with `NoSchedule`, but kube2iam must be scheduled on those nodes too, so a hardcoded `tolerations` block has been added to the `Deployment` resource. This is a short-term fix however, and `tolerations` should either be added to the chart values, or we should replace the use of our own chart with that of the kube2iam chart in the main chart repo.