cpanel 2.4.5: tweaks to worker livelinessProbe
Further tweaking liveliness probe for `worker` (Django Channels) container
to account for potentially very slow background tasks blocking a specific
worker.

Specifically new Deploy/Initialisation of RStudio can be VERY slow
and CP is doing a `helm upgrade --wait` so it is blocking the worker:

https://github.com/ministryofjustice/analytics-platform-control-panel/blob/main/controlpanel/frontend/consumers.py#L109-L132

The default `--timeout` for `helm` is 300s/5 minutes:

https://helm.sh/docs/helm/helm_upgrade/#options

So I'm tweaking the worker liveliness probe to consider a worker stalled when
it didn't run something for 5 minutes and reduce the chance that a worker
container is restarted just because someone installed a tool.

Part of ticket: https://trello.com/c/eUMRw1gK