cpanel 2.4.6: Moved `worker` in separate pod
The `worker` container is decoupled from the nginx/django frontend containers
and there is no reason for it to be in the same pod.

The workers are decoupled as they receive background jobs via redis.

Furthermore, have it in a separate `Deployment` means we can scale workers
independently from the frontend.

This also increase the number of workers replicas to `10` from `3`
(which was the number of frontend replicas).

I've added a `frontend.replicas` and `worker.replicas` values so that
these can potentially be tweaked/configured more easily.

Ticket: https://trello.com/c/Bbl5R56H