[ANPL-239] RStudio: Add an init container to change the user home on boot
We are using an init container as it will always finish no matter how long it takes.
Some user homes take a long time to chown as they contain a lot of conda packages.
We don't mind which group the home files are in as they can be both `users` and `staff` depending on which tool created it.